refactor: Convert LMAnime to multisrc
