+++ multisrc/overrides/animestream/lmanime/additional.gradle
@@ -1,0 +1,3 @@
+dependencies {
+    implementation(project(":lib-okru-extractor"))
+}
