+++ multisrc/overrides/animestream/lmanime/res/mipmap-hdpi/ic_launcher.png
