+++ multisrc/overrides/animestream/lmanime/res/mipmap-mdpi/ic_launcher.png
