+++ multisrc/overrides/animestream/lmanime/res/mipmap-xhdpi/ic_launcher.png
