+++ multisrc/overrides/animestream/lmanime/res/mipmap-xxhdpi/ic_launcher.png
