+++ multisrc/overrides/animestream/lmanime/res/mipmap-xxxhdpi/ic_launcher.png
