--- src/all/lmanime/build.gradle
@@ -1,19 +1,0 @@
-plugins {
-    alias(libs.plugins.android.application)
-    alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.kotlin.serialization)
-}
-
-ext {
-    extName = 'LMAnime'
-    pkgNameSuffix = 'all.lmanime'
-    extClass = '.LMAnime'
-    extVersionCode = 2
-}
-
-dependencies {
-    implementation(project(":lib-fembed-extractor"))
-    implementation(project(":lib-okru-extractor"))
-}
-
-apply from: "$rootDir/common.gradle"
